requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto==2.49.0
certifi==2023.7.22
colorama==0.4.6
colorlog==6.7.0
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
django-markwhat==1.6.2
djangorestframework==3.14.0
filelock==3.12.3
keyring==5.7.1
markdown2==2.4.10
numpy==1.25.2
pipenv==2023.9.8
platformdirs==3.10.0
psycopg2 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6064y7w0c6/croot/psycopg2_1687443467527/work
python-dotenv==1.0.0
pytz==2021.3
scipy==1.11.2
six==1.16.0
soupsieve==2.4.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tornado==6.3.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
virtualenv==20.24.5
